+++ src/main/java/service/member/MemberQuitService.java
@@ -73,7 +73,7 @@ public class MemberQuitService extends MemberBaseMapper {
         if(partyId!=null) criteria.andPartyIdEqualTo(partyId);
         if(branchId!=null) criteria.andBranchIdEqualTo(branchId);
 
-        return memberQuitMapper.countByExample(example);
+        return (int) memberQuitMapper.countByExample(example);
     }
 
     // 上一个 （查找比当前记录的“创建时间”  小  的记录中的  最大  的“创建时间”的记录）
